moved the saveTokens method to helpers
