+++ client/src/services/AuthService.ts
@@ -2,7 +2,7 @@ import { LoginResponseType, LoginType, User } from "../app/store/slice/auth/auth
 import { RegisterResponseType, RegisterType } from "../pages/Register/types";
 import { $api } from "../shared/config/api";
 import { routePath } from "../shared/config/routePath";
-import { saveTokensLocalStoare } from "../shared/config/saveTokens";
+import { saveTokensLocalStoare } from "../shared/helpers/saveTokens";
 
 class AuthService {
   static async login(userData: LoginType) {
